refactor: extract define_string_new_type macro to ozmux_macros crate
